[Fix] PDF 렌더 크기 타입 오류 수정
확대 배율에 따른 PDF 렌더 버킷 계산 결과가 Float로 추론되어 Int 크기 인자와 충돌하던 문제를 수정했습니다.

렌더 버킷은 정수 픽셀로 유지하고 배율 계산 구간에서만 Float 변환을 적용해 렌더 요청 타입을 일관되게 맞췄습니다.

Related to: #27

+++ NoteUp/app/src/main/java/com/kotlinsun/noteup/ui/canvas/CanvasFragment.kt
@@ -705,7 +705,8 @@ class CanvasFragment : Fragment() {
             return
         }
         val baseEdge = maxOf(binding.drawingCanvas.width, binding.drawingCanvas.height, 512)
-        val requestedEdge = ceil(baseEdge * scale / PDF_RENDER_BUCKET).toInt() * PDF_RENDER_BUCKET
+        val requestedEdge = ceil(baseEdge * scale / PDF_RENDER_BUCKET.toFloat()).toInt() *
+            PDF_RENDER_BUCKET
         val key = "${background.storageName}:${background.sourcePageIndex}:$requestedEdge"
         if (!force && key == pdfRenderKey) return
         pdfRenderKey = key
@@ -778,7 +779,7 @@ class CanvasFragment : Fragment() {
     }
 
     private companion object {
-        const val PDF_RENDER_BUCKET = 512f
+        const val PDF_RENDER_BUCKET = 512
         const val NOTE_ID_ARGUMENT = "noteId"
         const val INVALID_NOTE_ID = -1L
         const val DEFAULT_TEXT_WIDTH = 0.35f
